+++ quantforge/__init__.py
@@ -57,7 +57,9 @@ from .heston_mc import (
     heston_qe_mc, heston_cv_mc, heston_pathwise_delta, heston_mc_greeks,
 )
 from .bates import bates_price, bates_greeks, bates_smile
-from .double_heston import double_heston_price, double_heston_smile
+from .double_heston import (
+    double_heston_price, double_heston_greeks, double_heston_smile,
+)
 from .rough_heston import rough_heston_price, rough_heston_smile
 from .double_heston_calib import calibrate_double_heston
 from .heston_calib import calibrate_heston
@@ -261,7 +263,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.230.0"
+__version__ = "1.231.0"
 
 __all__ = [
     "OptionType",
@@ -333,6 +335,7 @@ __all__ = [
     "bates_greeks",
     "bates_smile",
     "double_heston_price",
+    "double_heston_greeks",
     "double_heston_smile",
     "rough_heston_price",
     "rough_heston_smile",